@@ -14,17 +14,18 @@ class CreateInsumoTable extends Migration
     public function up()
     {
         Schema::create('insumo', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->string('nombre')->nullable();
             $table->string('tipo');
             $table->text('composicion')->nullable();
             $table->string('envase')->nullable();
-            $table->float('existencias');
+            $table->decimal('existencias', 8, 2, true);
             $table->string('info')->nullable();
 
-            $table->unsignedInteger('unidad_medida_id');
+            $table->foreignId('unidad_medida_id');
             $table->foreign('unidad_medida_id')->references('id')->on('unidad_medida')->onDelete('cascade');
-            $table->unsignedInteger('subtipo_id');
+            
+            $table->foreignId('subtipo_id');
             $table->foreign('subtipo_id')->references('id')->on('subtipo')->onDelete('cascade');
         });
     }
