@@ -14,7 +14,7 @@ class CreateDestinatarioTable extends Migration
     public function up()
     {
         Schema::create('destinatario', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->string('nombre');
             $table->string('nucleo')->nullable();
         });
